refactor: prisma.service moved to services folder
